requirements.txt
requests
python-dotenv
uvicorn
fastapi
pydantic-settings
pydantic
pytest
azure-identity
azure-keyvault-secrets